ignore build output and add src
