+++ .gitignore
@@ -7,3 +7,5 @@ build
 lib/modules/
 lib/pypyjs.vm.*
 node_modules/
+lib/pypyjs.*
+lib/tests.*
