+++ src/tests/tests.js
@@ -3,8 +3,6 @@
 // We should do something a lot nicer than this...
 //
 
-'use strict';
-
 if (typeof vm === 'undefined') {
   if (typeof require !== 'undefined') {
     pypyjs = require('../pypyjs.js');
@@ -22,7 +20,7 @@ if (typeof console !== 'undefined') {
   log = print;
 }
 
-var toByteArray = function toByteArray(str) {
+var toByteArray = function(str) {
   var byteArray = [];
   for (var i = 0; i < str.length; i++) {
     if (str.charCodeAt(i) <= 0x7F) {
@@ -40,13 +38,13 @@ var toByteArray = function toByteArray(str) {
 var stdinBuffer = [];
 
 var vm = new pypyjs({
-  stdin: function stdin() {
+  stdin: function() {
     if (stdinBuffer.length > 0) {
       return stdinBuffer.pop();
     }
 
-    return new Promise(function (resolve) {
-      setTimeout(function () {
+    return new Promise(function(resolve) {
+      setTimeout(function() {
         var input = toByteArray('test\n');
         input.push(null);
         input.reverse();
@@ -60,12 +58,13 @@ var vm = new pypyjs({
 var pypyjsTestResult = vm.ready()
 
 // First, check that python-level errors will actually fail the tests.
-.then(function () {
+.then(function() {
   return vm.exec("raise ValueError(42)");
-}).then(function () {
+})
+.then(function() {
   throw new Error("Python exception did not trigger js Error");
-}, function (err) {
-  if (!err instanceof pypyjs.Error) {
+}, function(err) {
+  if (! err instanceof pypyjs.Error) {
     throw new Error("Python exception didn't trigger vm.Error instance");
   }
 
@@ -75,62 +74,128 @@ var pypyjsTestResult = vm.ready()
 })
 
 // Check that the basic set-exec-get cycle works correctly.
-.then(function () {
+.then(function() {
   return vm.set("x", 7);
-}).then(function () {
+})
+.then(function() {
   return vm.exec("x = x * 2");
-}).then(function () {
+})
+.then(function() {
   return vm.get("x");
-}).then(function (x) {
+})
+.then(function(x) {
   if (x !== 14) {
     throw new Error('set-exec-get cycle failed');
   }
 })
 
 // Check that eval() works correctly.
-.then(function () {
+.then(function() {
   return vm.eval("x + 1");
-}).then(function (x) {
+})
+.then(function(x) {
   if (x !== 15) {
     throw new Error('eval failed');
   }
-}).then(function () {
+})
+.then(function() {
   return vm.exec('from time import sleep; sleep(1)');
-}).then(function () {
-  var test = ['from time import sleep', 'def Test():', '    sleep(1)', '', 'Test()', 'class Test2:', '    def __init__(self):', '        sleep(1)', '    def sleepy(self):', '        sleep(1)', '', 'a = Test2()', 'a.sleepy()', '', 'usleep = lambda x: sleep(x/1000000.0)', 'usleep(100) #sleep during 100μs', ''].join('\r\n');
+}).then(function() {
+  var test = [
+    'from time import sleep',
+    'def Test():',
+    '    sleep(1)',
+    '',
+    'Test()',
+    'class Test2:',
+    '    def __init__(self):',
+    '        sleep(1)',
+    '    def sleepy(self):',
+    '        sleep(1)',
+    '',
+    'a = Test2()',
+    'a.sleepy()',
+    '',
+    'usleep = lambda x: sleep(x/1000000.0)',
+    'usleep(100) #sleep during 100μs',
+    '',
+  ].join('\r\n');
 
   return vm.exec(test);
-}).then(function () {
-  var test = ['from time import sleep', '', 'def sleep_decorator(function):', '', '    def wrapper(*args, **kwargs):', '        sleep(2)', '        return function(*args, **kwargs)', '    return wrapper', '', '@sleep_decorator', 'def print_number(num):', '    return num', '', 'print print_number(222)', '', 'for x in range(1,6):', '    print print_number(x)', ''].join('\r\n');
+}).then(function() {
+  var test = [
+    'from time import sleep',
+    '',
+    'def sleep_decorator(function):',
+    '',
+    '    def wrapper(*args, **kwargs):',
+    '        sleep(2)',
+    '        return function(*args, **kwargs)',
+    '    return wrapper',
+    '',
+    '@sleep_decorator',
+    'def print_number(num):',
+    '    return num',
+    '',
+    'print print_number(222)',
+    '',
+    'for x in range(1,6):',
+    '    print print_number(x)',
+    '',
+  ].join('\r\n');
 
   return vm.exec(test);
-}).then(function () {
-  gp = function () {
-    return new Promise(function (resolve, reject) {
-      setTimeout(function () {
+})
+.then(function() {
+  gp = function() {
+    return new Promise(function(resolve, reject) {
+      setTimeout(function() {
         resolve('test');
       }, 1000);
     });
   };
 
-  var test = ['import js', 'from time import time', '', 'print time()', "get_promise = js.eval('gp')", 'promise = get_promise()', 'res = js.await(promise)', 'print time()', 'print res', "assert res == 'test'"].join('\r\n');
+  var test = [
+    'import js',
+    'from time import time',
+    '',
+    'print time()',
+    "get_promise = js.eval('gp')",
+    'promise = get_promise()',
+    'res = js.await(promise)',
+    'print time()',
+    'print res',
+    "assert res == 'test'",
+  ].join('\r\n');
 
   return vm.exec(test);
-}).then(function () {
-  gp = function () {
-    return new Promise(function (resolve, reject) {
-      setTimeout(function () {
+})
+.then(function() {
+  gp = function() {
+    return new Promise(function(resolve, reject) {
+      setTimeout(function() {
         reject(new Error('test!'));
       }, 1000);
     });
   };
 
-  var test = ['import js', 'from time import time', '', 'print time()', "get_promise = js.eval('gp')", 'promise = get_promise()', 'res = js.await(promise)', 'print time()'].join('\r\n');
+  var test = [
+    'import js',
+    'from time import time',
+    '',
+    'print time()',
+    "get_promise = js.eval('gp')",
+    'promise = get_promise()',
+    'res = js.await(promise)',
+    'print time()',
+  ].join('\r\n');
 
   return vm.exec(test);
-}).then(function () {
+}).then(function() {
   throw new Error('Python exception did not trigger js Error');
-}, function (err) {
+},
+
+function(err) {
   if (!err instanceof pypyjs.Error) {
     throw new Error("Python exception didn't trigger vm.Error instance");
   }
@@ -142,35 +207,43 @@ var pypyjsTestResult = vm.ready()
 })
 
 // Check that we can read non-existent names and get 'undefined'
-.then(function () {
-  return vm.get("nonExistentName");
-}).then(function (x) {
+.then(function() {
+  return vm.get("nonExistentName")
+})
+.then(function(x) {
   if (typeof x !== 'undefined') {
     throw new Error('name should have been undefined');
   }
 })
 
 // Check that we execute in correctly-__name__'d python scope.
-.then(function () {
-  return vm.exec("assert __name__ == '__main__', __name__");
+.then(function() {
+  return vm.exec("assert __name__ == '__main__', __name__")
 })
 
 // Check that sys.platform tells us something sensible.
-.then(function () {
+.then(function() {
   return vm.exec("import sys; assert sys.platform == 'js'");
 })
 
 // Check that multi-line exec will work correctly.
-.then(function () {
+.then(function() {
   return vm.exec("x = 2\ny = x * 3");
-}).then(function () {
-  return vm.get("y");
-}).then(function (y) {
+})
+.then(function() {
+  return vm.get("y")
+})
+.then(function(y) {
   if (y !== 6) {
     throw new Error("multi-line exec didn't work");
   }
-}).then(function () {
-  var test = ['a = raw_input()', 'print a', "assert a == 'test'"].join('\r\n');
+})
+.then(function() {
+  var test = [
+    'a = raw_input()',
+    'print a',
+    "assert a == 'test'",
+  ].join('\r\n');
   return vm.exec(test);
 })
 
@@ -183,38 +256,42 @@ var pypyjsTestResult = vm.ready()
   });
 })
 // Check that multi-import statements will work correctly.
-.then(function () {
-  return vm.exec("import os\nimport time\nimport sys\nx=time.time()");
-}).then(function () {
-  return vm.get("x");
-}).then(function (x) {
+.then(function() {
+  return vm.exec("import os\nimport time\nimport sys\nx=time.time()")
+})
+.then(function() {
+  return vm.get("x")
+})
+.then(function(x) {
   if (!x) {
     throw new Error("multi-line import didn't work");
   }
 })
 
 // Check that you can create additional VMs using `new`
-.then(function () {
-  var vm2 = new pypyjs();
-  return vm2.exec("x = 17").then(function () {
-    return vm2.get("x");
-  }).then(function (x) {
+.then(function() {
+  var vm2 = new pypyjs()
+  return vm2.exec("x = 17").then(function() {
+    return vm2.get("x")
+  }).then(function(x) {
     if (x !== 17) {
-      throw new Error("newly-created VM didn't work right");
+      throw new Error("newly-created VM didn't work right")
     }
-  }).then(function () {
-    return vm2.get("y");
-  }).then(function (y) {
+  }).then(function() {
+    return vm2.get("y")
+  }).then(function(y) {
     if (typeof y !== "undefined") {
       throw new Error("name should have been undefined in new VM");
     }
-  });
+  })
 })
 
 // Report success or failure at the end of the chain.
-.then(function (res) {
+.then(function(res) {
   log('TESTS PASSED!');
-}, function (err) {
+},
+
+function(err) {
   log('TESTS FAILED!');
   log(err);
   throw err;
